pyproject.toml
[tool.poetry]
name = "ebird-telegram-bot"
version = "0.1.0"
description = ""
authors = ["andregri <andrea.grillo96@live.com>"]
readme = "README.md"
packages = [{include = "ebird_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
